Crate subspace_metrics

This Rust module serves as a bridge between two different Prometheus metrics libraries used: prometheus-client (official library) and TiKV’s prometheus client (used by Substrate). The module exposes a web server endpoint at “/metrics” that outputs metrics in Prometheus format. It adapts metrics from either or both of those libraries.

Enums

RegistryAdapter

Metrics registry adapter for prometheus-client and Substrate frameworks. It specifies which metrics registry or registries are in use.

Functions

start_prometheus_metrics_server

Start prometheus metrics server on the provided address.